Rename to game_controller_spl
