+++ README.md
@@ -1,4 +1,4 @@
-# GC SPL
+# Game Controller SPL
 
 ROS2 Package to interface RoboCup Standard Platform League's Game Controller.
