+++ game_controller_spl/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/game_controller_spl
+[install]
+install_scripts=$base/lib/game_controller_spl
