+++ game_controller_spl/test/test_copyright.py
