+++ game_controller_spl_interfaces/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>gc_spl_interfaces</name>
+  <name>game_controller_spl_interfaces</name>
   <version>0.0.0</version>
   <description>RoboCup SPL GameController Data ROS msg</description>
   <maintainer email="kenjibrameld@gmail.com">ijnek</maintainer>
