--- gc_spl/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/gc_spl
-[install]
-install_scripts=$base/lib/gc_spl
